+++ src/Resources/app/administration/src/provider/keycloak/heptacom-admin-open-auth-client-edit-page/index.js
@@ -1,0 +1,17 @@
+import template from './heptacom-admin-open-auth-client-edit-page.html.twig';
+
+const { Component } = Shopware;
+
+Component.override('heptacom-admin-open-auth-client-edit-page', {
+    template,
+
+    watch: {
+        item(newValue) {
+            if (newValue && newValue.provider === 'keycloak') {
+                if (!newValue.config.scopes) {
+                    newValue.config.scopes = [];
+                }
+            }
+        }
+    }
+});